==> gesture-analysis/scripts/07 Prepare data for modelling - initial feature selection.ipynb <==
# Prepare data for modelling: initial feature selection

Most algorithms in scikit learn can not handle a 7kx70k matrix as data. Therefor we use simple filters as a first step to reduce the numbers of fitlers.

%load_ext autoreload
%autoreload 2

import gc
import pickle
import pathlib
import tqdm
import numpy as np
import numpy.random
import pandas as pd
from sklearn.feature_selection import VarianceThreshold
from sklearn.feature_selection import SelectKBest, SelectPercentile
from sklearn.feature_selection import chi2, f_classif, mutual_info_classif
from sklearn.preprocessing import RobustScaler
from sklearn.model_selection import StratifiedShuffleSplit

base_path = "/home/jsimon/Documents/thesis/gesture-analysis/data/"
time_groups_path_corrected_pickl = base_path+"transformed/time_added/all/time-and-groups-corrected-all.pkl"
stats_added_base_path = base_path+"transformed/stats_added/all/"
stats_added_path_pickl = stats_added_base_path+"raw_stats-added-all.pkl"
gyro_calibration_path = base_path+'../scripts/gestureanalysis/gyro_offset.txt'

import os
os.getcwd()

with open( time_groups_path_corrected_pickl, "rb" ) as users_pickle_file:
    users = pickle.load(users_pickle_file)

with open( stats_added_base_path+'train.pkl', "rb" ) as users_pickle_file:
    traindata = pickle.load(users_pickle_file)

with open( stats_added_base_path+'validation.pkl', "rb" ) as users_pickle_file:
    validdata = pickle.load(users_pickle_file)

train_data_df = traindata['train']['data']
train_labels_df = traindata['train']['labels']

## Prepare the Training set

train_data_df.info()

# we fill the NaN values:
train_data_df.fillna(method='backfill', inplace=True)

headers = list(train_data_df.columns)
print(headers[0:3])

nh = len(headers)
headers = np.array(headers).reshape(1,nh)

X = train_data_df.values
print(X.shape)

np.argwhere(np.isnan(X)).shape

def simple_label_vec(labels_df, gestures):
    y = np.zeros((len(labels_df),))
    for index, row in labels_df.iterrows():
        if row.values.sum() == 0:
            y[index] = 0
        else:
            idx = row.idxmax()
            y[index] = gestures.index(idx) + 1
    return y

gestures = list(train_labels_df.columns.copy())
gestures.sort()
print(gestures)

y = simple_label_vec(train_labels_df, gestures)

y = y.astype(np.int)

y.shape

## Prepare the validation set

valid_data_df = validdata['valid']['data']
valid_labels_df = validdata['valid']['labels']

valid_data_df.fillna(method='backfill', inplace=True)

Xval = valid_data_df.values
print(Xval.shape)

yval = simple_label_vec(valid_labels_df, gestures)

yval = yval.astype(np.int)

yval.shape

## Scale the Data

transformer = RobustScaler().fit(X)
X = transformer.transform(X)
Xval = transformer.transform(Xval)

with open( stats_added_base_path+'scaler.pkl', "wb" ) as users_pickle_file:
    pickle.dump(transformer, users_pickle_file)

with open( stats_added_base_path+'scaler.pkl', "rb" ) as users_pickle_file:
    transformer = pickle.load(users_pickle_file)

## Clear Memory

del [[train_data_df,train_labels_df,valid_data_df,valid_labels_df]]
gc.collect()
transformer = None
traindata = None
validdata = None
train_data_df = None
train_labels_df = None
valid_data_df = None
valid_labels_df = None
gc.collect()
train_data_df = pd.DataFrame()
train_labels_df = pd.DataFrame()
valid_data_df = pd.DataFrame()
valid_labels_df = pd.DataFrame()

## Initial Feature selection

With 7201 features and > 70k instances our training set is to complex for most algorithms to fit well. Therfor I perform simple initial feature selection without a model to reduce the features who are f.e. only one value or clearly redundant and so on.

sel = VarianceThreshold().fit(X) # default: remove constants
X = sel.transform(X)
Xval = sel.transform(Xval)
print(headers.shape)
headers = sel.transform(headers)
print(headers.shape)

X.shape

def subsample(X,y,times):
    X_samples = X.copy()
    y_samples = y.copy()
    for i in range(times):
        sampler = StratifiedShuffleSplit(n_splits=2, test_size=0.5)
        sampler.get_n_splits(X_samples, y_samples)
        train_index, test_index = list(sampler.split(X_samples, y_samples))[0]
        X_samples = X_samples[train_index]
        y_samples = y_samples[train_index]
        gc.collect()
    return X_samples, y_samples

X_samples, y_samples = subsample(X,y,5)

X_samples.shape

sel = None
sampler = None
gc.collect()

sel = SelectPercentile(f_classif, percentile=60).fit(X_samples, y_samples) # keep a bit more than half of the features
candidate1 = sel.transform(X)
candidateVal1 = sel.transform(Xval)
candidateHeaders1 = sel.transform(headers)

candidate1.shape

sel = SelectPercentile(mutual_info_classif, percentile=60).fit(X_samples, y_samples) # keep a bit more than half of the features
candidate2 = sel.transform(X)
candidateVal2 = sel.transform(Xval)
candidateHeaders2 = sel.transform(headers)

candidate2.shape

with open( stats_added_base_path+'train-data-scaled-noconst.pkl', "wb" ) as users_pickle_file:
    pickle.dump({
        'X': X, 'y': y, 'Xval' : Xval, 'yval': yval, 
        'gestures' : gestures, 'headers': headers
    }, users_pickle_file)

with open( stats_added_base_path+'train-data-scaled-f-ANOVA60.pkl', "wb" ) as users_pickle_file:
    pickle.dump({
        'X': candidate1, 'y': y, 'Xval' : candidateVal1, 'yval': yval, 
        'gestures' : gestures, 'headers': candidateHeaders1
    }, users_pickle_file)

with open( stats_added_base_path+'train-data-scaled-mutual-inf60.pkl', "wb" ) as users_pickle_file:
    pickle.dump({
        'X': candidate2, 'y': y, 'Xval' : candidateVal2, 'yval': yval, 
        'gestures' : gestures, 'headers': candidateHeaders2
    }, users_pickle_file)
